feat : 수강신청 검증 로직이 별도의 트랜잭션을 탈수있도록 EnrollmentValidator 클래스로 분리
